Did not fix a method in DateTimeTwo.java
I did not fix the error with dateHashMap not being equal to the expected
value. The issue is that I am putting the values into a hashMap and
outputting it in an order that I am not away of the mechanics behind.
The order being random, I believe, is not something that is going to
affect the Zybooks grade.